Only install apache2 if not installed

+++ scripts/before_install.sh
@@ -5,8 +5,9 @@ if which yum &>/dev/null; then
   yum install -y httpd
   chkconfig httpd on
 elif which apt-get &>/dev/null; then
-  apt-get update
-  apt-get -y upgrade
-  apt-get -y install apache2
-  rm -f /var/www/html/index.html
+  if ! dpkg-query -l apache2 | grep -q "^ii"; then
+    apt-get update
+    apt-get -y upgrade
+    apt-get -y install apache2
+  fi
 fi
